Switch runtime to JRE base image and clean up Healthcheck class
- Runtime stage now uses sonatype-infosec/jre:openjdk-17 instead of
  the full JDK, reducing attack surface (no compiler in production)
- Rename healthcheck -> Healthcheck (public class, capitalized)
- Use 2-space indentation
- Fix --app mode to use response-code based exit status
